requirements.txt
streamlit
tensorflow
numpy
python-chess
scikit-learn
matplotlib
seaborn